some stuff modified

+++ src/main/java/com/hotel/java/application/services/PrecioService.java
@@ -1,5 +1,5 @@
 package com.hotel.java.application.services;
 
 public interface PrecioService {
-    int calculaDescuento(long dias);
+    double calculaDescuento(long dias);
 }
